+++ opensourcetest/__init__.py
@@ -1,2 +1,2 @@
-__version__ = '0.2.1'
+__version__ = '0.2.2'
 __description__ = "We need more free software interface testing."